# Conflicts:
#	src/components/grid/buttonDeleteComponent.js
#	src/components/grid/component.js
#	src/components/grid/constants.js
#	src/components/selectsComponent/constants.js
#	src/reducers/index.js

+++ src/components/grid/buttonDeleteComponent.js
@@ -6,7 +6,9 @@ import {deleteServer} from './actions';
 import {contactsByClientFindServer,clearContactCreate,clearContactOrder} from '../contact/actions';
 import {shareholdersByClientFindServer,clearShareholderCreate,clearShareholderOrder} from '../shareholder/actions';
 import {visitByClientFindServer} from '../visit/actions';
-import {NUMBER_RECORDS,DELETE_TYPE_CONTACT,DELETE_TYPE_SHAREHOLDER,DELETE_TYPE_VISIT} from './constants';
+import {DELETE_TYPE_VISIT,NUMBER_RECORDS,DELETE_TYPE_CONTACT,DELETE_TYPE_SHAREHOLDER, DELETE_PARTICIPANT_VIEW, DELETE_TASK_VIEW} from './constants';
+import {deleteParticipant} from '../participantsVisitPre/actions';
+import {deleteTask} from '../visit/createVisit/tasks/actions';
 
 class ButtonDeleteComponent extends Component{
 
@@ -21,18 +23,38 @@ class ButtonDeleteComponent extends Component{
      this._closeDelete = this._closeDelete.bind(this);
   }
 
-    _onConfirmDelete(){
-      const {actionsDelete,deleteServer} = this.props;
-      deleteServer(actionsDelete.urlServer,actionsDelete.json,actionsDelete.typeDelete).then((data) => {
-        if((_.get(data, 'payload.status') === 200)){
-            this.setState({showEx: true});
-            console.log(actionsDelete);
-          } else {
-            this.setState({showEr: true});
+  _onConfirmDelete(){
+    const {actionsDelete, deleteServer, participants, tasks} = this.props;
+    if(actionsDelete.typeDelete !== DELETE_PARTICIPANT_VIEW && actionsDelete.typeDelete !== DELETE_TASK_VIEW){
+          deleteServer(actionsDelete.urlServer,actionsDelete.json,actionsDelete.typeDelete).then((data) => {
+            if((_.get(data, 'payload.status') === 200)){
+                this.setState({showEx: true});
+                console.log(actionsDelete)
+              } else {
+                this.setState({showEr: true});
+            }
+            }, (reason) => {
+              this.setState({showEr: true});
+          });
+        }else {
+        if(actionsDelete.typeDelete === DELETE_PARTICIPANT_VIEW){
+          this.setState({show: false});
+          const {deleteParticipant} = this.props;
+          var indexDelete = participants.findIndex(function(item){
+            return item.idParticipante === actionsDelete.id;
+          });
+          deleteParticipant(indexDelete);
+        }else{
+          if(actionsDelete.typeDelete === DELETE_TASK_VIEW){
+            this.setState({show: false});
+            const {deleteTask} = this.props;
+            var indexDelete = tasks.findIndex(function(item){
+              return item.uuid === actionsDelete.id;
+            });
+            deleteTask(indexDelete);
+          }
         }
-        }, (reason) => {
-          this.setState({showEr: true});
-      });
+      }
     }
 
     _closeDelete(){
@@ -53,14 +75,14 @@ class ButtonDeleteComponent extends Component{
           else if(actionsDelete.typeDelete === DELETE_TYPE_VISIT){
             visitByClientFindServer(window.localStorage.getItem('idClientSelected'),0,NUMBER_RECORDS,"sh.sharePercentage",1,"");
           }
-        }
-        this.setState({showEx:false, showEr: false,show: false});
+      }
     }
 
-    render(){
-      const {actionsDelete,deleteGridReducer} = this.props;
-      return (
-      <td style={{padding: '10px', textAlign: 'center'}}>
+
+  render(){
+    const {actionsDelete,deleteGridReducer} = this.props;
+    return (
+    <td style={{padding: '10px', textAlign: 'center'}}>
       <button onClick={() => this.setState({ show: true })} className="btn btn-sm  btn-danger">
           <i style={{margin:'0em', fontSize : '1.2em'}} className="trash outline icon"></i>
         </button>
@@ -71,7 +93,7 @@ class ButtonDeleteComponent extends Component{
               confirmButtonColor= '#DD6B55'
               confirmButtonText= 'Sí, estoy seguro!'
               cancelButtonText = "Cancelar"
-              text={actionsDelete.mensaje != null  ? actionsDelete.mensaje : ''}
+              text={actionsDelete.mensaje}
               showCancelButton= {true}
               onCancel= {() => this.setState({show: false })}
               onConfirm={() => this._onConfirmDelete()}/>
@@ -101,13 +123,17 @@ function mapDispatchToProps(dispatch) {
   return bindActionCreators({
     deleteServer,contactsByClientFindServer,clearContactCreate,
     shareholdersByClientFindServer,clearShareholderCreate,clearShareholderOrder,clearContactOrder,
-    visitByClientFindServer
+    visitByClientFindServer,
+    deleteParticipant,
+    deleteTask
   }, dispatch);
 }
 
-function mapStateToProps({deleteGridReducer}, ownerProps) {
+function mapStateToProps({deleteGridReducer, participants, tasks}, ownerProps) {
   return {
-    deleteGridReducer
+    deleteGridReducer,
+    participants,
+    tasks
   };
 }
